Everywhere: Print bus cycle count in places the program often dies.
Not super useful for anything in particular, but it can give some
idea as to how fast the emulator is running. I'm measuring about
2.2 MHz on my Xeon W-2245. Will try to test on other CPUs soon like
Ryzen 9 5950X and Apple M1 Max.

In the future it would be cool to have some better timing mechanisms,
but this is good enough for now.