Backend+TUI: Allow loading of custom kernels directly.
The Kernel should be an ELF file and should load something to the reset vector
if it wants to get anywhere.

This method skips boot{0,1,2} and IOS and directly launches the Kernel.
Seems to work okay so far with the StarStruck Kernel. It doesn't get
super far, but it does get through _start just fine.