Backend+TUI: More work on Debug backend
Create a MPSC (we are not using multiple senders, but this is what Rust has for us built in) between the Debug backend
and the interp backend. The interpbackend will check it each CPU step and handle anything it gets.

It's stubbed out right now, but it will be able to read and write registers, read and write memory (DMA), and step the Cpu.