Backend: ARM LUT: Take the top bits of the opcode into account
The unconditional extentension space (top bits all 1s) uses these bits as part of the opcode, not as a condition.
We need to store it or there will be overlap with other instructions, causing mass confusion (ask me how I know)

This is super overkill and naive to do it like this, the compiler doesn't want to either, hence disabling
the const eval limits. It also wastes a ton of space, half a meg to be exact. Future me will find a way to
make this smaller while still being constant-evaluated.