Backend: Implement LdmXX with proper PC support
for exception handler returning.

Note: This currently checks if PC & 1 == 1 and sets thumb mode,
the ARM reference manual does *not* say to do this, it says to load
the PC and not touch it and let the SPSR determine if the next instruction will be
thumb or not. This may need to be corrected in the future, but it shouldn't interfere
with any thing either way.