Core: SHA: Multi-version SHA-1 impl to enable SIMD
This enables the compiler to auto-vectorize the SHA-1 algorithm for the
following CPU architectures / features:

x86/64
  - SSE4
  - AVX2 + FMA

aarch64
  - SVE2 + NEON
  - NEON

Other architectures will fall back to the scalar version which will be
optimized based on the rust compilers enabled target features.

Feature detection is done at runtime, but only once. After the first runtime
detection, the correct function pointer is stored in a OnceCell and used by
subsequent runs, minimizing the performance penalty for feature
detection.