Core: SHA: Silence a stupid clippy lint
This is midly annoying. Taking a reference to a const is not guaranteed
to work as expected (even though it does in this case). The clippy lint
advises to make it static instead of const, but in order to make it
static we need to impl Sync which is not allowed for a raw OnceCell due
to it containing a UnsafeCell. Of course we can not cross-crate impl
Sync for OnceCell, so we have to define our own little struct and impl
Sync for that. Even though it's immediately moved, we have to also add
!#[feature(const_mut_refs)] to have a const fn new() on our wrapper
struct. We are already using this feature in the backend, so no big
deal to add it to core as well.