Core: SDHC: Correct error reporting
This condition actually was possible.
Before we properly handled masked interrupts, this was not an issue, because it was
an accepted limitation of the emulator that interrupts that could not be serviced
synchronously(because they were masked) were dropped. Since this is now supported,
we have to be more explicit which conditions are actually an emulation error on our part.