Core: add / stub some more registers for Hollywood
In my experimentation with trying to get BC to launch, this was the
minimum of what was needed to get into the core of BC, and have it start
trying to launch MIOS.  Granted, it still fails trying to boot MIOS -
from my reverse engineering of it, it seems like the error code it's
flashing means that some SHA-1 hash failed to validate (or at least,
2 sets of 5 x u32s didn't compare identically, not sure what else
it'd be other than a SHA-1)).  But, this at least allows BC to boot
to some extent.  Maybe some future fix (like what's going on with
boot1d?) will end up fixing BC as well :)